+++ CHANGELOG.md
@@ -4,7 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.0] - 2018-10-25
+## [1.2.0] - 2018-12-03
+### Added
+ - REST JSON and XML-RPC users grants status column.
+ - Tested up to WordPress 5.0.0.
+
+### Changed
+ - Updated translations and screenshots.
+ - Documentation improvements.
+
+## [1.1.0] - 2018-11-25
 ### Added
  - Allows to remove REST API Really Simple Discovery (RSD) endpoint.
  - Allows to remove `<link>` to the Really Simple Discovery service endpoint on front-end pages.